smartest/requirements.txt
javalang==0.13.0
networkx==2.5.1
# gensim==3.8.2
smart_open==5.2.1
# scikit-learn==0.24.2
# scikit-learn==1.1.3
astor==0.8.1
tree-sitter==0.20.4
pyyaml==6.0.1
pymongo==4.1.1
GitPython==3.1.20
# Flask==2.0.3
Flask==2.2.5
Flask-Cors==3.0.10
# jinja2<3.1.0
# numpy<2.0
